+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [0.0.3] - SNAPSHOT
 
+### New
+- Get current status.
+
 ## [0.0.2] - 26 Jan 2024
 
 ### New